Adding tests for can show link
